pyproject.toml
[project]
name = "costco-analysis"
version = "0.1.0"
description = "Costco receipt analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "python-dateutil>=2.8.2",
    "python-dotenv>=1.0.0",
    "duckdb>=1.2.0",
]
